Major change to model theory
Instead of having a firm volume calculated for the cell wall case, the theory was revised so that a mechanically-induced turgor pressure pushes back against the osmotic pressure for the case of cell swelling. This updated theory allows for a direct solution of dVol/dt for all scenarios, and allowed for computation of the state space of the adapted system.