Created solvers for GRN from Graphs...x5
Added in an optimization-based solver to find equilibrium states in larger networks, along with a stability characterization routine for equilibria found. Also in the process of adding in network plotting routines, as well as plots for degree analysis. Finally, added in read-write capability so graph models can be written to file.