Moving away from GeneNetworks...x1
The probability networks model is much more elegant and easy to use and maintain than the original way we did this. Therefore, today a base class was defined and all useful methods from gene networks were moved into the base class, which is inherited by probability networks. This is a major change and the whole system is likely broken now...but we're working on it.