Working on Time Sims...x2
Implemented really cool functionality that allows the user to specify a starting state and a set of input signal "words" and to perform a time simulation where steady-states are automatically found from the time sim. This helps verify the State Transition Network and to visualize the actual system dynamics in time.