Working on Time Sims...x3
Implemented really cool functionality that makes a stability-glyph annotated plot of a time trajectory obtained from allowing the user to specify a starting state and a set of input signal "words" and to perform a time simulation where steady-states are automatically found from the time sim. This helps verify the State Transition Network and to visualize the actual system dynamics in time.