Working on Control Loops...x1
Back to working on the original control loop problem. Discovered lots of bugs while simulating, and have determined that the beta-constants need to be scaled by n_add_edges in order to show consistent behaviour between different systems. Otherwise, the simple control system works when there is a attractive limit cycle for all values of the environmental constant and no bifurcation of the system as the environmental constant changes. These systems benefit from *no* multistability.