Building a Boolean Network Solver...x3
Building in functionality so that the RNM works with a GRN described by a Boolean model. In this next step, we've modified the Boolean method so that it truly performs an "OR" function when considering multiple inputs (which was found to be necessary to match with continuous models). We also implemented the ability to constrain nodes as inputs. We're now ready to go onto build this into a Finite State Machine...