Building a Boolean Network Solver...x5
Building in functionality so that the RNM works with a GRN described by a Boolean model. In this next step, we've implemented the state transition constructor, which makes general and event-driven NFSM from a Boolean network model...also we have gotten rid of the normalizing denomenator in the additive boolean net model, while continuing to use "sign" to compute the "MAX" of the additive (OR) interaction.